+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add readyStates as both static and and instance members
 - default response `status` to 200
+- Pass mocked `ProgressEvent` and `MockEvent` to event listeners
 
 ## 1.0.0